refactor: move firebase functions to helpers
