+++ src/helpers/index.ts
@@ -1,1 +1,2 @@
 export * from './i18n';
+export * from './firebase';
